src/player_experience/api/requirements.txt
# TTA Player Experience API Requirements
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database drivers
redis==5.0.1
neo4j==5.15.0
psycopg2-binary==2.9.9

# Authentication and security
python-jose[cryptography]==3.3.0
# Authentication and password hashing
passlib==1.7.4
bcrypt==3.2.2  # Compatible with passlib 1.7.4 - provides __about__ module
python-multipart==0.0.6

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0

# WebSocket support
websockets==12.0

# Validation and utilities
email-validator==2.1.0
python-dateutil==2.8.2

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
